introduce error objects
until now errors were always string. this commit changes this to error
Objects, with the intention to be able to programmatically handle
errors better (without string parsing).

Example Use Case: Crawling a website and checking if links are valid.
Current State: when a link points to a ressource which is not
available, it returns an error with a string containing 404. State with
this commit, it returns an error with an errorType, statusCode etc,
which can be processed easily.