Disable osmosifying of raw functions
There is a "feature" that any function passed as one of properties of
object parameters to any command is replaced with than command. It is
rather inconvenience than handy, because it makes imposible to use
contextCallbacks in set command using object syntax. If setting several
simple properties can be bypassed, by calling set command several times,
using contextCallbacks for nested properties is completely imposible.

It is uncertain what was the reason behind such behaviour. It's hard for
me imagine where sole than command may be usefull in set or any oter
command parameter context. It neither documented, nor covered by tests,
so the proposal is to remove this "feature". This may cause some
regressions in some existing code that relies on such behaviour (is such
code exists), but it will be rather easy to fix by calling then commad
explisitly.

This commit also fixes #203 as `type: () => "some string"` now works.